fix bug pay deadline 1970 yang membuat data tidak terhapus, penggantian versi laravel ke laravel 8, update protect pembayaran untuk user yang sudah telat deadline di halaman payment, update modul mark delete status pada dashboard penggantian colum database competitor artinya harus migrate dulu. bug edit peserta, pembesaran ukuran logo sponsor utama

+++ resources/views/layouts/admin.blade.php
@@ -504,16 +504,18 @@
 
 <script>
 
-  @if (isset($paymentChecker))
+  @if (Auth::user()->isAdmin == 0)
+    @if (isset($paymentChecker))
 
-      Swal.fire({
-      title: 'Batas Waktu Pembayaran Berakhir',
-      html: "Pembayaran atas Nama <br> <strong>{{$paymentChecker['dataName']}}</strong> - <strong>{{$paymentChecker['dataDetail']->name . ' ' . $paymentChecker['level'] . $paymentChecker['fullGender'] }}</strong> dibatalkan karena batas waktu pembayaran telah berakhir. Silahkan melakukan pendaftaran ulang",
-      icon: 'error',
-      confirmButtonText: '<a class="text-white text-decoration-none" href="{{url('app/competitor')}}">Lanjutkan</a>'
-    })
-    {{App\Http\Controllers\PaymentChecker::deleteAutoLatePayment()}};
+        Swal.fire({
+        title: 'Batas Waktu Pembayaran Berakhir',
+        html: "Pembayaran atas Nama <br> <strong>{{$paymentChecker['dataName']}}</strong> - <strong>{{$paymentChecker['dataDetail']->name . ' ' . $paymentChecker['level'] . $paymentChecker['fullGender'] }}</strong> dibatalkan karena batas waktu pembayaran telah berakhir. Silahkan melakukan pendaftaran ulang",
+        icon: 'error',
+        confirmButtonText: '<a class="text-white text-decoration-none" href="{{url('app/competitor')}}">Lanjutkan</a>'
+      })
+      {{App\Http\Controllers\PaymentChecker::deleteAutoLatePayment()}};
 
+    @endif
   @endif
 
 </script>
